Fix issues in residency and multi-threaded fence tests
* Add fixes for residency tests
* Fix synchronization timeout in test_memory_multithread

+++ conformance_tests/core/test_multithread/src/test_memory_multithread.cpp
@@ -224,7 +224,7 @@ void perform_memory_manipulation() {
     lzt::append_memory_copy(cl, device_ptr, host_ptr, alloc_size, nullptr);
     lzt::close_command_list(cl);
     lzt::execute_command_lists(cq, 1, &cl, fence_);
-    lzt::sync_fence(fence_, UINT32_MAX);
+    lzt::sync_fence(fence_, UINT64_MAX);
     lzt::query_fence(fence_);
     lzt::reset_fence(fence_);
     EXPECT_EQ(ZE_RESULT_NOT_READY, zeFenceQueryStatus(fence_));
